wkkk

+++ baiano.js
@@ -2015,23 +2015,25 @@ case 'sticker':
       }
        break
 
-        case 's':
-        if (!isGroup) return
-        if (isMedia && !mek.message.videoMessage || isQuotedImage) {
-      const encmedia = isQuotedImage ? JSON.parse(JSON.stringify(mek).replace('quotedM', 'm')).message.extendedTextMessage.contextInfo : mek
-      const media = await client.downloadAndSaveMediaMessage(encmedia, `./${senderfix}`)
-      reply('please wait')
-      await ffmpeg(`${media}`)
-      .input(media)
-      .on('start', function (cmd) {
-      console.log(`Started : ${cmd}`)
-      })
-      .on('error', function (err) {
-      console.log(`Error : ${err}`)
-      fs.unlinkSync(media)
-      reply('error')
-        })
-      .on('end', async function () {
+        case 'stf':
+				if (!isUser) return registroA()
+   
+					
+				if ((isMedia && !mek.message.videoMessage || isQuotedImage) && args.length == 0) {
+						const encmedia = isQuotedImage ? JSON.parse(JSON.stringify(mek).replace('quotedM','m')).message.extendedTextMessage.contextInfo : mek
+						const media = await client.downloadAndSaveMediaMessage(encmedia)
+						ran = getRandom('.webp')
+						await ffmpeg(`./${media}`)
+							.input(media)
+							.on('start', function (cmd) {
+								console.log(`Started : ${cmd}`)
+							})
+							.on('error', function (err) {
+								console.log(`Error : ${err}`)
+								fs.unlinkSync(media)
+								reply(mess.stikga)
+							})
+							.on('end', async function () {
       console.log('Finish')
       const MatadataFix3 = {
         type: 'full',
@@ -2041,29 +2043,32 @@ case 'sticker':
             '🌹'
         ]
         }
-      const sticker4s = await createSticker(`./${senderfix}.webp`, MatadataFix3)
-      client.sendMessage(from, sticker4s, MessageType.sticker, {quoted: mek})
-      fs.unlinkSync(media)  
-      fs.unlinkSync(`./${senderfix}.webp`)  
-      })
-      .addOutputOptions([`-vcodec`,`libwebp`,`-vf`,`scale='min(320,iw)':min'(320,ih)':force_original_aspect_ratio=decrease,fps=15, pad=320:320:-1:-1:color=white@0.0, split [a][b]; [a] palettegen=reserve_transparent=on:transparency_color=ffffff [p]; [b][p] paletteuse`])
-      .toFormat('webp')
-      .save(`./sticker/${senderfix}.webp`)
-      } else if ((isMedia && mek.message.videoMessage.fileLength < 10000000 || isQuotedVideo && mek.message.extendedTextMessage.contextInfo.quotedMessage.videoMessage.fileLength < 10000000)) {
-      const encmedia = isQuotedVideo ? JSON.parse(JSON.stringify(mek).replace('quotedM', 'm')).message.extendedTextMessage.contextInfo : mek
-      const media = await client.downloadAndSaveMediaMessage(encmedia, `./${senderfix}`)
-      reply('❬❗❭ *_Esᴘᴇʀᴀ ᴇɴ ᴘʀᴏᴄᴇsᴏ.._*')
-      await ffmpeg(`${media}`)
-      .inputFormat(media.split('.')[4])
-      .on('start', function (cmd) {
-      console.log(`Started : ${cmd}`)
-      })
-      .on('error', function (err) {
-      console.log(`Error : ${err}`)
-      fs.unlinkSync(media)
-      tipe = media.endsWith('.mp4') ? 'video' : 'gif'
-      })
-      .on('end', async function () {
+        buffer = fs.readFileSync(ran)
+      const sticker4s = await createSticker(buffer, MatadataFix3)
+								client.sendMessage(from, sticker4s, sticker, {quoted: mek})
+								fs.unlinkSync(media)
+								fs.unlinkSync(ran)
+							})
+							.addOutputOptions([`-vcodec`,`libwebp`,`-vf`,`scale='min(320,iw)':min'(320,ih)':force_original_aspect_ratio=decrease,fps=15, pad=320:320:-1:-1:color=white@0.0, split [a][b]; [a] palettegen=reserve_transparent=on:transparency_color=ffffff [p]; [b][p] paletteuse`])
+							.toFormat('webp')
+							.save(ran)
+					} else if ((isMedia && mek.message.videoMessage.seconds < 10 || isQuotedVideo && mek.message.extendedTextMessage.contextInfo.quotedMessage.videoMessage.seconds < 10) && args.length == 0) {
+						const encmedia = isQuotedVideo ? JSON.parse(JSON.stringify(mek).replace('quotedM','m')).message.extendedTextMessage.contextInfo : mek
+						const media = await client.downloadAndSaveMediaMessage(encmedia)
+						ran = getRandom('.webp')
+						reply(mess.wait) 
+						await ffmpeg(`./${media}`)
+							.inputFormat(media.split('.')[1])
+							.on('start', function (cmd) {
+								console.log(`Started : ${cmd}`)
+							})
+							.on('error', function (err) {
+								console.log(`Error : ${err}`)
+								fs.unlinkSync(media)
+								tipe = media.endsWith('.mp4') ? 'video' : 'gif'
+								reply(mess.stikga)
+							})
+							.on('end', async function () {
       console.log('Finish')
       const MatadataFix4 = {
         type: 'full',
@@ -2073,18 +2078,20 @@ case 'sticker':
             '🌹'
         ]
         }
-      const sticker5s = await createSticker(`./${senderfix}.webp`, MatadataFix4)
+       buffer = fs.readFileSync(ran)
+      const sticker5s = await createSticker(buffer, MatadataFix4)
       client.sendMessage(from, sticker5s, MessageType.sticker, {quoted: mek})
-      fs.unlinkSync(media)
-      fs.unlinkSync(`./${senderfix}.webp`)
-      })
-      .addOutputOptions([`-vcodec`,`libwebp`,`-vf`,`scale='min(320,iw)':min'(320,ih)':force_original_aspect_ratio=decrease,fps=15, pad=320:320:-1:-1:color=white@0.0, split [a][b]; [a] palettegen=reserve_transparent=on:transparency_color=ffffff [p]; [b][p] paletteuse`])
-      .toFormat('webp')
-      .save(`./${senderfix}.webp`)
-          } else {
-      reply(`please mention a picture`)
-      }
-        break
+								fs.unlinkSync(media)
+								fs.unlinkSync(ran)
+							})
+							.addOutputOptions([`-vcodec`,`libwebp`,`-vf`,`scale='min(320,iw)':min'(320,ih)':force_original_aspect_ratio=decrease,fps=15, pad=320:320:-1:-1:color=white@0.0, split [a][b]; [a] palettegen=reserve_transparent=on:transparency_color=ffffff [p]; [b][p] paletteuse`])
+							.toFormat('webp')
+							.save(ran)
+							} else {
+						reply(`Enviar imagem / vídeo / gif com legenda \n${prefix}sticker (duração do adesivo de vídeo de 1 a 9 segundos)`)
+					} 
+					
+					break
 
 
 /*case 'stickerf': //by gauger
